Let members choose short or long hair, without blush or flecks
The avatar builder had no hair choice because only the long-hair sheet
was ready. Luke's short-hair sheet is now finished, painted in the same
colors as the long one, so the palette swap works on it unchanged.

- characterSheetPath picks long.png or short.png by the avatar's
  hairStyle for both the world and AvatarPreview; no saved avatar still
  means long.
- The builder gets a Short / Long choice above Hair Color, and new
  members start with long hair (DEFAULT_AVATAR).
- A data migration moves every saved "short" to "long". The builder
  saved "short" by default while the world only drew long hair, so
  members chose their colors against long hair; nobody's look changes.
- The cheek blush is painted out of both runtime sheets (art, not code,
  so it's gone even for the default look).
- Avatars in privacy-hardened browsers had pink and purple flecks: Luke's
  Firefox perturbs canvas reads (about 12% of pixels one step off), so an
  exact color lookup left those pixels in the sheet's original paint.
  recolorPixels now snaps each pixel to the nearest sheet color within 2
  steps (the closest paints are 10 apart), snaps alpha to 0 or 255, and
  recolorSheet throws on a read that is mostly noise so the loader falls
  back to the sheet's own colors instead of baking static.

README gains v0.10.4; CREDITS notes the modified sheets; the art README
records where each runtime sheet comes from; the world rules describe
the canvas noise and how to reproduce it.

+++ src/lib/types/game.ts
@@ -43,7 +43,7 @@ export const PANTS_COLORS = ["#353540", "#1e3a5f", "#3b2219", "#4a5568", "#2d374
 export const SHOES_COLORS = ["#4d3f38", "#353540", "#ca5954", "#ede4da", "#1e3a5f"] as const;
 
 export const DEFAULT_AVATAR: AvatarConfig = {
-  hairStyle: "short",
+  hairStyle: "long",
   hairColor: "#3b2219",
   skinTone: "#C68642",
   shirtColor: "#5c699f",
